[added] type check
