+++ src/database/type.ts
@@ -1,0 +1,4 @@
+export type Like = {
+  id: number;
+  contentId: string;
+};
